Moved SquadReader into a reading_comprehension/ module, removed old sentence selection code
