+++ allennlp/data/dataset_readers/__init__.py
@@ -8,7 +8,7 @@ and produce datasets in the formats required by specific models.
 
 from allennlp.data.dataset_readers.dataset_reader import DatasetReader
 from allennlp.data.dataset_readers.language_modeling import LanguageModelingReader
+from allennlp.data.dataset_readers.reading_comprehension import SquadReader
 from allennlp.data.dataset_readers.sequence_tagging import SequenceTaggingDatasetReader
 from allennlp.data.dataset_readers.snli import SnliReader
-from allennlp.data.dataset_readers.squad import SquadReader, SquadSentenceSelectionReader
 from allennlp.data.dataset_readers.semantic_role_labeling import SrlReader
