+++ allennlp/data/dataset_readers/reading_comprehension/__init__.py
@@ -1,0 +1,6 @@
+"""
+Reading comprehension is loosely defined as follows: given a question and a passage of text that
+contains the answer, answer the question.
+"""
+
+from allennlp.data.dataset_readers.reading_comprehension.squad import SquadReader
